rename game screen file
