+++ js/render-screen.js
@@ -1,7 +1,7 @@
 import getIntro from './screen-intro';
-import getGreeting from './screen-greeting';
+import getGreeting from '../screen-greeting';
 import getRuls from './screen-rules';
-import getGame from './game-screen';
+import getGame from './screen-game';
 import getStats from './screen-stats';
 
 const container = document.querySelector(`.central`);
